Extract validation deps logic
