+++ crates/holochain/src/core/queue_consumer/sys_validation_consumer.rs
@@ -2,8 +2,8 @@
 
 use super::*;
 use crate::core::workflow::sys_validation_workflow::sys_validation_workflow;
+use crate::core::workflow::sys_validation_workflow::validation_deps::ValidationDependencies;
 use crate::core::workflow::sys_validation_workflow::SysValidationWorkspace;
-use crate::core::workflow::sys_validation_workflow::ValidationDependencies;
 use crate::core::IncomingDhtOpSender;
 use parking_lot::Mutex;
 use tracing::*;
@@ -22,11 +22,6 @@ pub fn spawn_sys_validation_consumer(
     let workspace = Arc::new(workspace);
     let space = Arc::new(space);
 
-    // Create an incoming ops sender for any dependencies we find
-    // that we are meant to be holding but aren't.
-    // If we are not holding them they will be added to our incoming ops.
-
-    // TODO circular import, move somewhere else
     let current_validation_dependencies = Arc::new(Mutex::new(ValidationDependencies::new()));
 
     super::queue_consumer_dna_bound(
@@ -35,6 +30,9 @@ pub fn spawn_sys_validation_consumer(
         conductor.task_manager(),
         (tx.clone(), rx),
         move || {
+            // Create an incoming ops sender for any dependencies we find
+            // that we are meant to be holding but aren't.
+            // If we are not holding them they will be added to our incoming ops.
             let incoming_dht_ops_sender =
                 IncomingDhtOpSender::new(space.clone(), trigger_self.clone());
             sys_validation_workflow(
